Refactor field translations for clarity
