+++ fmm/src/lib.rs
@@ -10,6 +10,9 @@ pub mod types;
 mod field_translation {
     pub mod hadamard;
     pub mod source;
-    pub mod source_to_target;
+    pub mod source_to_target {
+        pub mod fft;
+        pub mod svd;
+    }
     pub mod target;
 }
